pyproject.toml
[tool.poetry]
name = "drf-sample"
version = "0.1.0"
description = ""
authors = ["k4ssyi <kasikasi00@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.2"
djangorestframework = "^3.10"
django-filter = "^2.2"
django-environ = "^0.4.5"
drf-writable-nested = "^0.5.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.7"
isort = "^4.3"
pyformat = "^0.7.0"
bpython = "^0.18.0"
django-debug-toolbar = "^2.0"
django-extensions = "^2.2"
drf-extensions = "^0.5.0"
Werkzeug = "^0.15.5"
django-cors-headers = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
